+++ app/src/main/java/com/inky/fitnesscalendar/db/Database.kt
@@ -29,7 +29,7 @@ import com.inky.fitnesscalendar.util.DATABASE_NAME
 private const val TAG = "Database"
 
 @Database(
-    version = 35,
+    version = 36,
     entities = [
         Activity::class,
         Recording::class,
@@ -64,6 +64,7 @@ private const val TAG = "Database"
         AutoMigration(from = 32, to = 33),
         AutoMigration(from = 33, to = 34),
         AutoMigration(from = 34, to = 35),
+        AutoMigration(from = 35, to = 36, spec = Migration35To36Spec::class),
     ]
 )
 @TypeConverters(Converters::class)