+++ app/src/main/java/com/inky/fitnesscalendar/di/DecisionTrees.kt
@@ -9,7 +9,6 @@ import com.inky.fitnesscalendar.db.entities.RichActivity
 import com.inky.fitnesscalendar.util.decision_tree.DecisionTree
 import com.inky.fitnesscalendar.util.decision_tree.Example
 import com.inky.fitnesscalendar.util.decision_tree.Examples
-import com.inky.fitnesscalendar.util.getCurrentBssid
 import com.inky.fitnesscalendar.util.toLocalDateTime
 import kotlinx.parcelize.Parcelize
 import java.time.LocalDateTime
@@ -45,8 +44,7 @@ object DecisionTrees {
         activities: List<RichActivity>,
     ): DecisionTree<ActivityType> {
         val examples = Examples(activities.map {
-            val attributes =
-                attributes(it.activity.startTime.toLocalDateTime(), it.activity.wifiBssid)
+            val attributes = attributes(it.activity.startTime.toLocalDateTime())
             Example(it.type, attributes)
         })
 
@@ -57,8 +55,7 @@ object DecisionTrees {
         activities: List<RichActivity>,
     ): DecisionTree<Vehicle> {
         val examples = Examples(activities.map {
-            val attributes =
-                attributes(it.activity.startTime.toLocalDateTime(), it.activity.wifiBssid)
+            val attributes = attributes(it.activity.startTime.toLocalDateTime())
             Example(it.activity.vehicle, attributes)
         })
 
@@ -67,15 +64,14 @@ object DecisionTrees {
 
     private fun learnPlace(activities: List<RichActivity>): DecisionTree<Place> {
         val examples = Examples(activities.map {
-            val attributes =
-                attributes(it.activity.startTime.toLocalDateTime(), it.activity.wifiBssid)
+            val attributes = attributes(it.activity.startTime.toLocalDateTime())
             Example(it.place, attributes)
         })
 
         return DecisionTree.learn(examples)
     }
 
-    private fun attributes(date: LocalDateTime, wifiBssid: String?): List<Any?> {
+    private fun attributes(date: LocalDateTime): List<Any?> {
         val hourOfDay = date.toLocalTime().hour
         // Segments:
         // 0: [2-6) Uhr
@@ -87,9 +83,9 @@ object DecisionTrees {
         val timeOfDay =
             ((22.0 + hourOfDay.toDouble()).mod(24.0) / 4).toInt()
         val weekDay = date.dayOfWeek
-        return listOf(timeOfDay, weekDay, wifiBssid)
+        return listOf(timeOfDay, weekDay)
     }
 
     private fun <T : Any> DecisionTree<T>.classifyNow(context: Context) =
-        classify(attributes(LocalDateTime.now(), context.getCurrentBssid()))
+        classify(attributes(LocalDateTime.now()))
 }