+++ app/src/main/java/com/inky/fitnesscalendar/ui/views/settings/SettingsView.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Menu
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -21,15 +20,11 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
@@ -37,8 +32,6 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.rememberMultiplePermissionsState
 import com.inky.fitnesscalendar.BuildConfig
 import com.inky.fitnesscalendar.R
 import com.inky.fitnesscalendar.preferences.Preference
@@ -115,8 +108,6 @@ fun SettingsView(
                     modifier = Modifier.padding(bottom = 8.dp)
                 )
 
-                CollectBssidPreference()
-
                 PreferenceToggle(Preference.PREF_ENABLE_PUBLIC_API)
 
                 Setting(
@@ -139,59 +130,6 @@ fun SettingsView(
                 }
             }
         }
-    }
-}
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-private fun CollectBssidPreference() {
-    var showRequestExplanationDialog by rememberSaveable { mutableStateOf(false) }
-    var tryingToEnablePreference by rememberSaveable { mutableStateOf(false) }
-
-    val context = LocalContext.current
-    val scope = rememberCoroutineScope()
-    val permissionsState = rememberMultiplePermissionsState(
-        permissions = listOf(
-            android.Manifest.permission.ACCESS_FINE_LOCATION,
-            android.Manifest.permission.ACCESS_COARSE_LOCATION
-        ),
-        onPermissionsResult = { result ->
-            if (tryingToEnablePreference && result.values.all { it }) {
-                tryingToEnablePreference = false
-                scope.launch {
-                    Preference.COLLECT_BSSID.set(context, true)
-                }
-            }
-        }
-    )
-
-    PreferenceToggle(preference = Preference.COLLECT_BSSID) { enabled ->
-        tryingToEnablePreference = false
-        if (enabled && !permissionsState.allPermissionsGranted) {
-            showRequestExplanationDialog = true
-            false
-        } else {
-            true
-        }
-    }
-
-    if (showRequestExplanationDialog) {
-        AlertDialog(
-            onDismissRequest = {
-                showRequestExplanationDialog = false
-            },
-            confirmButton = {
-                TextButton(onClick = {
-                    showRequestExplanationDialog = false
-                    tryingToEnablePreference = true
-                    permissionsState.launchMultiplePermissionRequest()
-                }) {
-                    Text(stringResource(R.string.action_continue))
-                }
-            },
-            title = { Text(stringResource(R.string.wifi_information_permission_help_title)) },
-            text = { Text(stringResource(R.string.wifi_information_permission_help_description)) }
-        )
     }
 }