+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Remove "Store location information" preference
+
 # 2025.04.1
 
 - Add a public api that allows third party apps to create activities